pyproject.toml
[project]
name = "margin-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "google-auth>=2.40.3",
    "gspread-asyncio>=2.0.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
]
